# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "aimem hub API",
    "version": "v1",
    "description": "The complete HTTP surface of an aimem hub (and of the local service, which serves the same mux over a unix socket without auth). Every route except the three public pages (/, /v1/status, /admin) requires `Authorization: Bearer <token>`; routes marked x-role admin additionally need an admin token (the env token, or a tokens.json entry with role admin). This spec documents the surface and its auth contract; request/response bodies are intentionally loose (JSON objects) - the store types in internal/store are the authority on field shapes. A parity test (openapi_test.go) fails CI when this file and the real route table disagree."
  },
  "servers": [{"url": "https://hub.example.com:8440"}],
  "components": {
    "securitySchemes": {
      "bearer": {"type": "http", "scheme": "bearer"}
    },
    "parameters": {
      "p": {"name": "p", "in": "path", "required": true, "schema": {"type": "string"}, "description": "project id"},
...
      "name": {"name": "name", "in": "path", "required": true, "schema": {"type": "string"}, "description": "document or provider name"}
    }
  },
  "security": [{"bearer": []}],
  "paths": {
...
    "/v1/projects/{p}/docs": {
      "get": {"summary": "List shared documents (name, rev, writer; tombstones flagged)", "x-role": "writer", "parameters": [{"$ref": "#/components/parameters/p"}], "responses": {"200": {"description": "{docs}"}}}
    },
    "/v1/projects/{p}/docs/{name}": {
      "get": {"summary": "Read a shared document (current, or ?rev= for a retained revision)", "x-role": "writer", "parameters": [{"$ref": "#/components/parameters/p"}, {"$ref": "#/components/parameters/name"}], "responses": {"200": {"description": "document"}}},
      "put": {"summary": "Compare-and-swap write ({body, base_rev, updated_by}); named tokens prefix updated_by; identical body is a no-op", "x-role": "writer", "parameters": [{"$ref": "#/components/parameters/p"}, {"$ref": "#/components/parameters/name"}], "responses": {"200": {"description": "{rev}"}, "409": {"description": "stale base_rev - carries current rev, writer, and (capped) body to merge from"}}},
      "delete": {"summary": "Tombstone a document (?base_rev=&by=; CAS like any write)", "x-role": "writer", "parameters": [{"$ref": "#/components/parameters/p"}, {"$ref": "#/components/parameters/name"}], "responses": {"200": {"description": "{rev, deleted}"}, "409": {"description": "stale base_rev"}}}
    },
...
  }
}